requirements.txt
undetected-chromedriver==3.1.7
mysql-connector-python==8.0.28
selenium==4.1.3